feat(api): routes and static class refactor
